+++ app.js
@@ -6,12 +6,7 @@ const { PORT = 3000 } = process.env;
 
 const app = express();
 
-app.use('/', routes);
-// app.use('*', (req, res) => {
-//   res.status(404);
-//   res.send({ message: 'Запрашиваемый ресурс не найден' });
-// });
-
 app.use(express.static(path.join(__dirname, 'public')));
+app.use('/', routes);
 
 app.listen(PORT);
